made canvas.platform.ts which you need to make an gameobject. In maps.ts the objects are being configured. And little things moved around
