+++ src/app/helpers/canvas.character.ts
@@ -1,12 +1,6 @@
 import { MovingObject } from './canvas.moving.object';
 
 export class Character extends MovingObject{
-  public update() {
-    this.newPos();
-    this.ctx.fillStyle = this.color;
-    this.ctx.fillRect(this.x, this.y, this.width, this.height);
-  }
-
   public moveup() {
     if (this.y >= 600) {
       this.gravitySpeed = -15;
@@ -26,8 +20,4 @@ export class Character extends MovingObject{
     this.speedrightX = +3;
   }
 
-  public stop() {
-    this.speedleftX = 0;
-    this.speedrightX = 0;
-  }
 }
