+++ src/app/helpers/canvas.platform.ts
@@ -1,0 +1,7 @@
+import {CanvasObject} from './canvas.object';
+
+export class CanvasPlatform extends CanvasObject {
+  constructor(ctx, width, height, x, y, color){
+    super(ctx, width, height, x, y, color);
+  }
+}
